Delete the now-dead marker helpers and extract the shared ScopedEnvironmentVariable
Delegating RunningUnderContinuousIntegration() to the shared header left
ContinuousIntegrationMarkers() and AnyEnvironmentVariableIsSet() in
FrozenBaselinePolicy.{h,cc} with zero callers; deletes both, and the now-
unused <span> include. Verified by grep that neither PR 1095's shader
validators nor PR 1096's Metal device gate call them. PR 1095
(shader-validators-fail-closed-on-ci) deletes the same two functions on its
own branch, so this converges with it rather than conflicting.

Also extracts donner/base/tests/ScopedEnvironmentVariable.h, taken byte-for-
byte from that branch, and migrates every local copy of the class onto it:
FrozenBaselinePolicy_tests.cc (matching that branch's own migration of the
same file), ContinuousIntegrationMarkers_tests.cc, and
EnvironmentCapabilityGate_tests.cc. Three copies down to one.
