+++ donner/base/BUILD.bazel
@@ -128,6 +128,7 @@ donner_cc_library(
         "tests/EnvironmentCapabilityGate.h",
         "tests/ParseResultTestUtils.h",
         "tests/Runfiles.h",
+        "tests/ScopedEnvironmentVariable.h",
         "tests/TestTempDir.h",
     ],
     # //tools tests (MCP servers, build tooling) share the same test-hygiene
