+++ donner/gpu/baseline/BUILD.bazel
@@ -42,6 +42,7 @@ donner_cc_test(
     srcs = ["FrozenBaselinePolicy_tests.cc"],
     deps = [
         ":frozen_baseline_policy",
+        "//donner/base:base_test_utils",
         "@com_google_gtest//:gtest_main",
     ],
 )
